pyproject.toml
[project]
name = "ai-lod"
version = "0.1.0"
description = "hybrid lod simplification with qem + ai importance"
authors = [{name = "pran"}]
requires-python = ">=3.9"
dependencies = [
    "trimesh>=4.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pyrender>=0.1.45",
    "pyglet<2",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
ai-lod = "ai_lod.cli.simplify:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
